refactor: Add `escape` file, add as export
Reduces the need for importing the index of the module further.
