+++ package.json
@@ -27,6 +27,10 @@
         "./lib/decode.js": {
             "require": "./lib/decode.js",
             "import": "./lib/esm/decode.js"
+        },
+        "./lib/escape.js": {
+            "require": "./lib/escape.js",
+            "import": "./lib/esm/escape.js"
         }
     },
     "files": [
