+++ src/encode.ts
@@ -1,50 +1,7 @@
-import { encodeHTMLTrieRe, getCodePoint } from "./encode-trie.js";
+import htmlTrie from "./generated/encode-html.js";
+import { xmlReplacer, getCodePoint } from "./escape.js";
 
 const htmlReplacer = /[\t\n!-,./:-@[-`\f{-}$\x80-\uFFFF]/g;
-const xmlReplacer = /["&'<>$\x80-\uFFFF]/g;
-
-const xmlCodeMap = new Map([
-    [34, "&quot;"],
-    [38, "&amp;"],
-    [39, "&apos;"],
-    [60, "&lt;"],
-    [62, "&gt;"],
-]);
-
-/**
- * Encodes all non-ASCII characters, as well as characters not valid in XML
- * documents using XML entities.
- *
- * If a character has no equivalent entity, a
- * numeric hexadecimal reference (eg. `&#xfc;`) will be used.
- */
-export function encodeXML(str: string): string {
-    let ret = "";
-    let lastIdx = 0;
-    let match;
-
-    while ((match = xmlReplacer.exec(str)) !== null) {
-        const i = match.index;
-        const char = str.charCodeAt(i);
-        const next = xmlCodeMap.get(char);
-
-        if (next !== undefined) {
-            ret += str.substring(lastIdx, i) + next;
-            lastIdx = i + 1;
-        } else {
-            ret += `${str.substring(lastIdx, i)}&#x${getCodePoint(
-                str,
-                i
-            ).toString(16)};`;
-            // Increase by 1 if we have a surrogate pair
-            lastIdx = xmlReplacer.lastIndex += Number(
-                (char & 0b1111_1111_1000_0000) === 0xd800
-            );
-        }
-    }
-
-    return ret + str.substr(lastIdx);
-}
 
 /**
  * Encodes all entities and non-ASCII characters in the input.
@@ -70,78 +27,49 @@ export function encodeNonAsciiHTML(data: string): string {
     return encodeHTMLTrieRe(xmlReplacer, data);
 }
 
-/**
- * Encodes all non-ASCII characters, as well as characters not valid in XML
- * documents using numeric hexadecimal reference (eg. `&#xfc;`).
- *
- * Have a look at `escapeUTF8` if you want a more concise output at the expense
- * of reduced transportability.
- *
- * @param data String to escape.
- */
-export const escape = encodeXML;
+function encodeHTMLTrieRe(regExp: RegExp, str: string): string {
+    let ret = "";
+    let lastIdx = 0;
+    let match;
 
-function getEscaper(
-    regex: RegExp,
-    map: Map<number, string>
-): (data: string) => string {
-    return function escape(data: string): string {
-        let match;
-        let lastIdx = 0;
-        let result = "";
+    while ((match = regExp.exec(str)) !== null) {
+        const i = match.index;
+        ret += str.substring(lastIdx, i);
+        const char = str.charCodeAt(i);
+        let next = htmlTrie.get(char);
 
-        while ((match = regex.exec(data))) {
-            if (lastIdx !== match.index) {
-                result += data.substring(lastIdx, match.index);
+        if (typeof next === "object") {
+            // We are in a branch. Try to match the next char.
+            if (i + 1 < str.length) {
+                const nextChar = str.charCodeAt(i + 1);
+                const value =
+                    typeof next.n === "number"
+                        ? next.n === nextChar
+                            ? next.o
+                            : undefined
+                        : next.n.get(nextChar);
+
+                if (value !== undefined) {
+                    ret += value;
+                    lastIdx = regExp.lastIndex += 1;
+                    continue;
+                }
             }
 
-            // We know that this chararcter will be in the map.
-            result += map.get(match[0].charCodeAt(0))!;
-
-            // Every match will be of length 1
-            lastIdx = match.index + 1;
+            next = next.v;
         }
 
-        return result + data.substring(lastIdx);
-    };
+        // We might have a tree node without a value; skip and use a numeric entitiy.
+        if (next !== undefined) {
+            ret += next;
+            lastIdx = i + 1;
+        } else {
+            const cp = getCodePoint(str, i);
+            ret += `&#x${cp.toString(16)};`;
+            // Increase by 1 if we have a surrogate pair
+            lastIdx = regExp.lastIndex += Number(cp !== char);
+        }
+    }
+
+    return ret + str.substr(lastIdx);
 }
-
-/**
- * Encodes all characters not valid in XML documents using XML entities.
- *
- * Note that the output will be character-set dependent.
- *
- * @param data String to escape.
- */
-export const escapeUTF8 = getEscaper(/[&<>'"]/g, xmlCodeMap);
-
-/**
- * Encodes all characters that have to be escaped in HTML attributes,
- * following {@link https://html.spec.whatwg.org/multipage/parsing.html#escapingString}.
- *
- * @param data String to escape.
- */
-export const escapeAttribute = getEscaper(
-    /["&\u00A0]/g,
-    new Map([
-        [34, "&quot;"],
-        [38, "&amp;"],
-        [160, "&nbsp;"],
-    ])
-);
-
-/**
- * Encodes all characters that have to be escaped in HTML text,
- * following {@link https://html.spec.whatwg.org/multipage/parsing.html#escapingString}.
- *
- * @param data String to escape.
- */
-export const escapeText = getEscaper(
-    /[&<>\u00A0]/g,
-    new Map([
-        [38, "&amp;"],
-        [60, "&lt;"],
-        [62, "&gt;"],
-        [160, "&nbsp;"],
-    ])
-);
